+++ action.go
@@ -4,23 +4,6 @@ import (
 	"sync"
 )
 
-// HandleAction handles the named action with the given handler.
-func HandleAction(name string, h ActionHandler) {
-	actions.Handle(name, h)
-}
-
-// PostAction creates and posts the named action with the given arg.
-// The action is handled in its own goroutine.
-func PostAction(name string, arg interface{}) {
-	actions.Post(name, arg)
-}
-
-// PostActions creates and posts a batch of actions.
-// All the actions are handled sequentially in a separate goroutine.
-func PostActions(a ...Action) {
-	actions.PostBatch(a...)
-}
-
 // Action represents an action to handle.
 type Action struct {
 	Name string
