+++ action_test.go
@@ -1,21 +1,43 @@
-package app_test
+package app
 
-// func TestActions(t *testing.T) {
-// 	app.Handle("test", func(e app.EventDispatcher, a app.Action) {})
+import (
+	"sync"
+	"testing"
+)
 
-// 	app.NewAction("test", 42)
-// 	app.NewActions(
-// 		app.Action{Name: "test", Arg: 21},
-// 		app.Action{Name: "test", Arg: 84},
-// 	)
-// }
+func TestActions(t *testing.T) {
+	HandleAction("test", func(e EventDispatcher, a Action) {})
 
-// func TestActionRegistry(t *testing.T) {
-// 	tests.TestActionRegistry(t, func() app.ActionRegistry {
-// 		dispatcher := app.NewEventRegistry(func(f func()) {
-// 			f()
-// 		})
+	PostAction("test", 42)
+	PostActions(
+		Action{Name: "test", Arg: 21},
+		Action{Name: "test", Arg: 84},
+	)
+}
 
-// 		return app.NewActionRegistry(dispatcher)
-// 	})
-// }
+func TestActionRegistry(t *testing.T) {
+	var wg sync.WaitGroup
+
+	d := newEventRegistry(func(f func()) {
+		f()
+	})
+	r := newActionRegistry(d)
+
+	r.Handle("test", func(e EventDispatcher, a Action) {
+		wg.Done()
+	})
+
+	r.Post("unknown", nil)
+
+	wg.Add(1)
+	r.Post("test", 42)
+
+	wg.Add(3)
+	r.PostBatch(
+		Action{Name: "test", Arg: nil},
+		Action{Name: "test", Arg: "hello"},
+		Action{Name: "test", Arg: 21},
+	)
+
+	wg.Wait()
+}
