+++ app.go
@@ -14,16 +14,9 @@ var (
 
 	driver  Driver
 	factory = NewFactory()
-	actions *actionRegistry
-)
-
-func init() {
-	events := NewEventRegistry(CallOnUIGoroutine)
-	events = ConcurrentEventRegistry(events)
-	DefaultEventRegistry = events
-
+	events  = newEventRegistry(CallOnUIGoroutine)
 	actions = newActionRegistry(events)
-}
+)
 
 // Import imports the component into the app.
 // Components must be imported in order the be used by the app package.
@@ -171,4 +164,29 @@ func Dock() DockTile {
 // UI goroutine is the running application main thread.
 func CallOnUIGoroutine(f func()) {
 	driver.CallOnUIGoroutine(f)
+}
+
+// HandleAction handles the named action with the given handler.
+func HandleAction(name string, h ActionHandler) {
+	actions.Handle(name, h)
+}
+
+// PostAction creates and posts the named action with the given arg.
+// The action is handled in its own goroutine.
+func PostAction(name string, arg interface{}) {
+	actions.Post(name, arg)
+}
+
+// PostActions creates and posts a batch of actions.
+// All the actions are handled sequentially in a separate goroutine.
+func PostActions(a ...Action) {
+	actions.PostBatch(a...)
+}
+
+// NewEventSubscriber creates an event subscriber to return when
+// implementing the app.Subscriber interface.
+func NewEventSubscriber() *EventSubscriber {
+	return &EventSubscriber{
+		registry: events,
+	}
 }
